Populate project_input_template folder with public data This now contains validation data for all feature DW branches, with the exception of Erection branch for which validation is still failing. This expected validation data was created using following steps: 1. Obtained expected validation data from Q4, FY19 deliverable results 2. Ran simulation using proprietar input data, and ensured validation passed 3. Replaced proprietary input data with public input data and re ran validation 4. Re-populated expected validation spreadsheet with results obtained from step 3. using purely public input data.
Once Erection modules validation data and code is corrected and the expected validation spreadsheet is updated,
the expected validation spreadsheet will be complete officially and will be used for validation moving forward.
That is, this expected validation spreadsheet contains accurate expected results for DW and utility scenarios
(with the exception of Erection data).